feat: :sparkles: create layout routes
