+++ src/app/layout.tsx
@@ -1,17 +1,24 @@
-import './globals.css'
+import { cookies } from 'next/headers'
 import { ReactNode } from 'react'
+import './globals.css'
+
+import Copyright from '@/components/Copyright'
+import Hero from '@/components/Hero'
+import Profile from '@/components/Profile'
+import SignIn from '@/components/SignIn'
+
 import {
-  Roboto_Flex as Roboto,
   Bai_Jamjuree as BaiJamjuree,
+  Roboto_Flex as Roboto,
 } from 'next/font/google'
 
 const roboto = Roboto({ subsets: ['latin'], variable: '--font-roboto' })
-
 const baiJamjuree = BaiJamjuree({
   subsets: ['latin'],
   weight: '700',
   variable: '--font-bai-jamjuree',
 })
+
 export const metadata = {
   title: 'NLW | Spacetime',
   description:
@@ -19,12 +26,33 @@ export const metadata = {
 }
 
 export default function RootLayout({ children }: { children: ReactNode }) {
+  const isAuthenticated = cookies().has('token-spacetime')
+
   return (
     <html lang="pt-BR">
       <body
         className={`${roboto.variable} ${baiJamjuree.variable} bg-gray-900 font-sans text-gray-100`}
       >
-        {children}
+        <main className="grid min-h-screen grid-cols-2">
+          {/* Left section */}
+          <div className="relative flex flex-col items-start justify-between  overflow-hidden border-r border-white/10 bg-[url(../assets/bg-stars.svg)] bg-cover px-28 py-16 ">
+            {/* Blur */}
+            <div className="absolute right-0 top-1/2 h-[288px] w-[526px] -translate-y-1/2 translate-x-1/2 rounded-full bg-purple-700 opacity-50 blur-full" />
+
+            {/* Stripes */}
+            <div className="absolute bottom-0 right-2 top-0 w-2 bg-stripes " />
+
+            {isAuthenticated ? <Profile /> : <SignIn />}
+
+            <Hero />
+            <Copyright />
+          </div>
+
+          {/* Right section */}
+          <div className="flex flex-col bg-[url(../assets/bg-stars.svg)] bg-cover p-16">
+            {children}
+          </div>
+        </main>
       </body>
     </html>
   )
