+++ src/app/memories/new/page.tsx
@@ -1,0 +1,7 @@
+export default function New() {
+  return (
+    <div>
+      <h1>New</h1>
+    </div>
+  )
+}
